parser/requirements.txt
requests
bs4
lxml
fake_useragent
mysql-connector-python